*V2* BIG CHANGE SPLIT PLANNER INTO many files +++
