+++ client/html/components/footer.html
@@ -1,4 +1,4 @@
-<template name="aboutFooter">
+<template name="footer">
 	<div class="container-fluid bg-primary footertext">
 		<p class="footer-content text-center text-secondary">© 2018 ItinPlan All Rights Reserved</p>
 	</div>
