+++ client/html/components/locationModalTemplate.html
@@ -1,0 +1,36 @@
+<template name="locationModalTemplate">
+    <!-- Local Modal-->
+    <!--modal for opening location @yuanrong javascript must be in planner template-->
+    <div class="modal fade" id="locationModal" tabindex="-1" role="dialog" aria-labelledby="modal display"
+         aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered" role="document">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h5 class="modal-title" id="exampleModalLabel">Select Location:</h5>
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <div class="modal-body">
+                    <form>
+                        <div class="form-row align-items-center">
+                            <div class="col-auto">
+                                <input type="text" class="form-control" id="recipient-name">
+                            </div>
+                            <div class="col-auto">
+                                <button type="button" class="btn btn-secondary btn-searchLoc">Search</button>
+                            </div>
+                        </div>
+
+                    </form>
+                    <div class="map-container">
+                        {{> googleMap name="locMap" options=locationMap}}
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-secondary btn-saveLoc" data-dismiss="modal">Save</button>
+                </div>
+            </div>
+        </div>
+    </div>
+</template>
