+++ client/html/components/locationTemplate.html
@@ -1,0 +1,30 @@
+<!-- template for each location -->
+<template name="locationTemplate">
+    <!-- New layout-->
+    <li class="timeline-line"></li>
+    <li class="timeline-item">
+        <div class="timeline-badge"><a href="#"></a></div>
+        <div class="timeline-panel">
+            <div class="timeline-heading">
+                {{getLocName}}
+            </div>
+            <div class="timeline-content">
+                <ul>
+                    <li>INFORMATION IN HERE</li>
+                    <li>PICTURE IN HERE</li>
+                </ul>
+            </div>
+            {{#if isOwner}}
+            <div class="timeline-footer">
+                <button class="close text-danger btn-deleteLoc">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+
+                <button class="btn btn-success btn-selectLoc" data-toggle="modal"
+                        data-target="#locationModal" data-location={{location}}>Select Location
+                </button>
+            </div>
+            {{/if}}
+        </div>
+    </li>
+</template>
