+++ client/html/layout/accounts.html
@@ -1,5 +1,5 @@
 <template name="account">
-    <section class="jumbotron loginback jumbotron-fluid">
+    <section class="jumbotron loginback jumbotron-fluid mb-0">
         {{> ensureSignedIn template="accountTemplate" auth="loginTemplate"}}
     </section>
 </template>
