+++ client/html/layout/home.html
@@ -8,8 +8,8 @@
                 <a class="btn btn-success" id="plan-btn" href="/planner/?_id=new">Plan New Trip</a>
             </div>
             {{#if haveLocalTrip}}
-                <div class="container text-center">
-                    <h7 style="color:wheat;">Or resume planning an <a style="color:lightskyblue;" href="/planner">older trip</a>...</h7>
+                <div class="container text-center pt-1">
+                    <h7 style="color:wheat;">Or resume planning an <a style="color:#7ff4ff;" href="/planner">older trip</a>...</h7>
                 </div>
             {{/if}}
         </div>
