+++ client/html/main.html
@@ -10,6 +10,6 @@
     {{> Template.dynamic template=content}}
 </body>
 <footer>
-	{{> aboutFooter}}
+	{{> footer}}
 </footer>
 </template>
