+++ client/js/locationTemplate.js
@@ -1,0 +1,44 @@
+/*
+	locationTemplate helper and events
+*/
+
+Template.locationTemplate.helpers({
+	//set this index to ind+1
+	getLocNum: function(index) {
+		return index+1;
+	},
+	//get location from db, return location name.
+	locationName: function() {
+		return this;
+	},
+	getLocName() {
+		var result = ReactiveMethod.call('getLocName', this.location);
+		return result.data.result.name;
+	}
+});
+
+Template.locationTemplate.events({
+	//remove this location from trip.
+	'click .btn-deleteLoc' (event) {
+		this.trip.get().dayArray[this.dayIndex].splice(this.locIndex, 1);
+		this.trip.set(this.trip.get());
+	},
+	//open the location modal
+	'click .btn-selectLoc' (event) {
+		dayIndex = this.dayIndex;
+		locIndex = this.locIndex;
+		//open a javascript modal thing
+		//gotten from bootstrap https://getbootstrap.com/docs/4.0/components/modal/?#varying-modal-content
+		$('#locationModal').one('show.bs.modal', function (event) {
+			var button = $(event.relatedTarget); // Button that triggered the modal
+			var location = button.data('location'); // Extract info from data-* attributes
+			// If necessary, you could initiate an AJAX request here (and then do the updating in a callback).
+			// Update the modal's content. We'll use jQuery here, but you could use a data binding library or other methods instead.
+			var modal = $(this);
+			modal.find('.modal-body input').val(location);
+			//set row and column for later setting data
+			modal.data("row", dayIndex);
+			modal.data("col", locIndex);
+		})
+	}
+});
