+++ scss/pages/_home.scss
@@ -2,7 +2,7 @@
   
   //sizing
   width: 100vm;
-  //height: 100vh;
+  height: 100vh;
   //margin: 0px;
   padding-top: 35vh;
   padding-bottom: 35vh;
